==> 4-combined-dataset-compilation.ipynb <==
# Combines the Mixed and Twitter data sets to create the training and test sets for a combined data set
import pandas as pd
import os

mix_training_df = pd.read_csv('emotion-labeled-data/mix/mix-training_t5.tsv', sep='\t').astype(str)
mix_testing_df = pd.read_csv('emotion-labeled-data/mix/mix-testing_t5.tsv', sep='\t').astype(str)

twit_training_df = pd.read_csv('emotion-labeled-data/twit0.825/twit0.825-training_t5.tsv', sep='\t').astype(str)
twit_testing_df = pd.read_csv('emotion-labeled-data/twit0.825/twit0.825-testing_t5.tsv', sep='\t').astype(str)

combo_training_df = pd.concat([mix_training_df, twit_training_df])
combo_testing_df = pd.concat([mix_testing_df, twit_testing_df])
combo_training_df = combo_training_df.drop(['Unnamed: 0'], axis=1)
combo_testing_df = combo_testing_df.drop(['Unnamed: 0'], axis=1)

# Ensures the directory exists for saving the data sets
path = 'emotion-labeled-data/combined'
if not os.path.exists(path):
    os.mkdir(path)

# Saves combined data set ready for the fine-tuning process
combo_training_df.to_csv('emotion-labeled-data/combined/combined-training_t5.tsv', sep='\t')
combo_testing_df.to_csv('emotion-labeled-data/combined/combined-testing_t5.tsv', sep='\t')
